base: exception method implementation moved to cpp
