+++ containers/test/CMakeLists.txt
@@ -2,7 +2,7 @@ set( KIT containers )
 file( GLOB source ${SOURCE_CODE_BASE_DIR}/${KIT}/test/*test.cpp )
 set( test_name ${CMAKE_PROJECT_NAME}_test_${KIT} )
 add_executable( ${test_name} ${source} )
-target_link_libraries( ${test_name} ${GTEST_BOTH_LIBRARIES} pthread ) # target_link_libraries( ${test_name} comma_${KIT} ${GTEST_BOTH_LIBRARIES} pthread )
+target_link_libraries( ${test_name} ${GTEST_BOTH_LIBRARIES} comma_base pthread ) # target_link_libraries( ${test_name} comma_${KIT} ${GTEST_BOTH_LIBRARIES} pthread )
 add_test( NAME ${test_name} COMMAND ${CMAKE_PROJECT_NAME}_test_${KIT} WORKING_DIRECTORY ${PROJECT_BINARY_DIR}/bin )
 if( INSTALL_TESTS )
     install( TARGETS ${test_name} RUNTIME DESTINATION ${comma_CPP_TESTS_INSTALL_DIR} COMPONENT Runtime )
